tests: tpm2d: Switch swtpm to --daemon
swtpm takes a bit of time to initialize and it might not be ready by the
time the next command is issued when we start it in background.

Fortunately, it has a --daemon parameter that allows to run in the
background but only returns once it is ready, so let's use it instead.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>