tests: tpm2d: Use PID file for SWTPM
The start_sw_tpm.sh script will setup the TPM to run a test. Its only
output is the swtpm or tpm_server PID so that the runner knows what
process to kill when the test is done.

We currently retrieve the PID by using the $! variable since we start
both processes in background.

However, it can lead to race conditions with swtpm where it wouldn't be
initialized yet by the time we run the next command, so we'll want to
remove the background job in favor of its --daemon argument.

We would still need to retrieve the PID somehow, but fortunately swtpm
provides the --pid option to store its PID in a file of our choosing.
Let's use that so we can switch to --deamon in a later patch.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>