configure: tpm: Run tests if we have either swtpm or tpm_server
The current logic will only compile (and thus run) the tests only if we
have either tpm_server, or if we have swtpm and tssstartup.

tsssstartup however is part of IBM's TSS, while we also support Intel's.

This effectively prevents the tests from running on top of Intel's TSS,
so let's relax the condition a bit and run the tests if we have either
swtpm or tpm_server.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>